+++ examples/spawn.rs
@@ -137,7 +137,6 @@ fn setup(
     let effect1 = effects.add(
         EffectAsset::new(vec![32768], Spawner::rate(500.0.into()), writer1.finish())
             .with_name("emit:rate")
-            .with_property("my_accel", Vec3::new(0., -3., 0.).into())
             .init(init_pos1)
             // Make spawned particles move away from the emitter origin
             .init(init_vel1)
@@ -254,7 +253,8 @@ fn setup(
     let init_size3 = SetAttributeModifier::new(Attribute::SIZE, size3);
 
     // Add property-driven acceleration
-    let accel3 = writer3.prop("my_accel").expr();
+    let my_accel = writer3.add_property("my_accel", Vec3::new(0., -3., 0.).into());
+    let accel3 = writer3.prop(my_accel).expr();
     let update_accel3 = AccelModifier::new(accel3);
 
     let init_pos3 = SetPositionSphereModifier {
@@ -275,7 +275,6 @@ fn setup(
             writer3.finish(),
         )
         .with_name("emit:burst")
-        .with_property("my_accel", Vec3::new(0., -3., 0.).into())
         .init(init_pos3)
         .init(init_vel3)
         .init(init_age3)